move files into Decoder and Encoder folder
