+++ Sources/RSocketCore/Extensions/Coder/Encoder/Encoder+CompositeMetadata.swift
