+++ Sources/RSocketCore/Extensions/Coder/Encoder/Encoder+MapData.swift
