+++ Sources/RSocketCore/Extensions/Coder/Encoder/Encoder+MapMetadata.swift
