+++ Sources/RSocketCore/Extensions/Coder/Encoder/Encoder+MetadataEncoder.swift
