+++ Sources/RSocketCore/Extensions/Coder/Encoder/Encoder+StaticMetadataEncoder.swift
